data_acqusition_tool/requirements.txt
matplotlib==3.4.3
pyserial==3.5
tk==0.1.0
